feat: add get interface type function
